# Conflicts:
#	cannon/cannon.iml
#	cannon/pom.xml
#	cannon/target/classes/cannongame/App.class
#	cannon/target/classes/cannongame/CalculateYValue.class
#	cannon/target/classes/cannongame/DegreeConversion.class
#	cannon/target/classes/cannongame/GettingFinalXCoordinate.class
#	cannon/target/classes/cannongame/IAngleV.class
#	cannon/target/classes/cannongame/IAttempt.class
#	cannon/target/classes/cannongame/IDegreeConversion.class
#	cannon/target/classes/cannongame/IGettingFinalXCoordinate.class
#	cannon/target/test-classes/cannongame/AngleTest.class
#	cannon/target/test-classes/cannongame/DegreeConversionTest.class
#	cannon/target/test-classes/cannongame/GettingFinalXCoordinateTest.class

+++ cannon/src/main/java/cannongame/GettingFinalXCoordinate.java
@@ -3,7 +3,7 @@ package cannongame;
 public class GettingFinalXCoordinate implements IGettingFinalXCoordinate {
 
 
-    public double gettingFinalXCoordinate(double degree, int velocity) {
+    public double gettingFinalXCoordinate(double degree, double velocity) {
 
         return Math.cos(degree) * velocity;
